+++ R/dashboardLabel.R
@@ -4,6 +4,7 @@
 #'
 #' @param ... any text.
 #' @param status label status: "danger", "success", "info", "primary", "warning".
+#' @param style labebl border style: "default" (rounded angles), "circled" or "square".
 #'
 #' @author David Granjon, \email{dgranjon@@ymail.com}
 #'
@@ -16,7 +17,9 @@
 #'     dashboardHeader(),
 #'     dashboardSidebar(),
 #'     dashboardBody(
-#'      dashboardLabel("Label 1", status = "info")
+#'      dashboardLabel("Label 1", status = "info"),
+#'      dashboardLabel("Label 2", status = "danger", style = "circle"),
+#'      dashboardLabel("Label 3", status = "success", style = "square")
 #'     )
 #'   ),
 #'   server = function(input, output) { }
@@ -24,10 +27,15 @@
 #' }
 #'
 #' @export
-dashboardLabel <- function(..., status = "primary") {
+dashboardLabel <- function(..., status = "primary", style = "default") {
   validateStatus(status)
   shiny::tags$span(
     class = paste0("label", " label-", status),
+    style = if (style == "circle") {
+      "border-radius: 50%;"
+    } else if (style == "square") {
+      "border-radius: 0%;"
+    } else NULL,
     ...
   )
 }